requirements.txt
antiorm==1.2.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
db==0.1.1
db-sqlite3==0.0.1
fake-useragent==0.1.11
gevent==20.6.2
greenlet==0.4.16
grequests==0.6.0
idna==2.9
lxml==4.5.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==4.0.0a6.post2
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
zope.event==4.4
zope.interface==5.1.0

mysql-connector-python~=8.0.20
